requirements.txt
jira==3.5.2
python-dotenv==1.0.0
ollama==0.6.1
chromadb==1.5.0
sentence-transformers==5.2.3
beautifulsoup4==4.12.3
requests==2.31.0

